requirements.txt
alembic==1.11.1
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
decorator==5.1.1
executing==1.2.0
Faker==18.11.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
ipdb==0.13.13
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
SQLAlchemy-serializer==1.4.1
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
